+++ src/native_dumper/dumper.py
@@ -1,18 +1,26 @@
-from collections import OrderedDict
 from gc import collect
 from io import (
     BufferedReader,
     BufferedWriter,
 )
 from logging import Logger
-from types import MethodType
 from typing import (
     Any,
     BinaryIO,
     Iterable,
+    NoReturn,
     Union,
 )
 
+from base_dumper import (
+    BaseDumper,
+    DBMetadata,
+    DumperMode,
+    IsolationLevel,
+    multiquery,
+    timeouts,
+    transfer_diagram,
+)
 from light_compressor import (
     CompressionMethod,
     auto_detector,
@@ -23,54 +31,60 @@ from nativelib import (
     NativeReader,
     NativeWriter,
 )
-from pandas import DataFrame as PdFrame
-from polars import DataFrame as PlFrame
-from sqlparse import format as sql_format
 
 from .common import (
     CHUNK_SIZE,
-    DBMS_DEFAULT_TIMEOUT_SEC,
     CHConnector,
     ClickhouseServerError,
-    DBMetadata,
-    DumperLogger,
     HTTPCursor,
     NativeDumperError,
     NativeDumperReadError,
     NativeDumperValueError,
     NativeDumperWriteError,
-    chunk_query,
     file_writer,
     make_columns,
-    transfer_diagram,
 )
+from .version import __version__
 
 
-class NativeDumper:
-    """Class for read and write Native format."""
+class NativeDumper(BaseDumper):
+    """Class for read and write Clickhouse Native format."""
+
+    cursor: HTTPCursor
 
     def __init__(
         self,
         connector: CHConnector,
         compression_method: CompressionMethod = CompressionMethod.ZSTD,
         logger: Logger | None = None,
-        timeout: int = DBMS_DEFAULT_TIMEOUT_SEC,
+        timeout: int | None = None,
+        isolation: IsolationLevel = IsolationLevel.committed,
+        mode: DumperMode = DumperMode.PROD,
     ) -> None:
         """Class initialization."""
 
-        if not logger:
-            logger = DumperLogger()
+
+        if int(connector.port) == 9000:
+            raise NativeDumperValueError(
+                "NativeDumper don't support port 9000, please, use 8123."
+            )
+
+        if timeout is None:
+            timeout = timeouts.CLICKHOUSE_DEFAULT_TIMEOUT
+
+        self.version = __version__
+        self.stream_type = "native"
+
+        super().__init__(
+            connector,
+            compression_method,
+            logger,
+            timeout,
+            isolation,
+            mode,
+        )
 
         try:
-            self.connector = connector
-
-            if int(self.connector.port) == 9000:
-                raise ValueError(
-                    "NativeDumper don't support port 9000, please, use 8123."
-                )
-
-            self.compression_method = compression_method
-            self.logger = logger
             self.cursor = HTTPCursor(
                 connector=self.connector,
                 compression_method=self.compression_method,
@@ -92,64 +106,31 @@ class NativeDumper:
             f"[{self.dbname} {self.version}]"
         )
 
-    @staticmethod
-    def multiquery(dump_method: MethodType):
-        """Multiquery decorator."""
+    @property
+    def timeout(self) -> int:
+        """Property method for get session timeout."""
 
-        def wrapper(*args, **kwargs):
+        return self._timeout
 
-            first_part: list[str]
-            second_part: list[str]
+    @timeout.setter
+    def timeout(self, timeout_value: int) -> int:
+        """Property method for set statement_timeout."""
 
-            self: NativeDumper = args[0]
-            cursor: HTTPCursor = (kwargs.get("dumper_src") or self).cursor
-            query: str = kwargs.get("query_src") or kwargs.get("query")
-            part: int = 1
-            first_part, second_part = chunk_query(self.query_formatter(query))
-            total_parts = len(sum((first_part, second_part), [])) + int(
-                bool(kwargs.get("table_name") or kwargs.get("table_src"))
-            )
+        self._timeout = timeout_value
+        self.cursor.timeout = self._timeout
+        return self._timeout
 
-            if len(first_part) > 1:
-                for query in first_part:
-                    self.logger.info(f"Execute query {part}/{total_parts}")
-                    cursor.execute(query)
-                    part += 1
+    @property
+    def isolation(self) -> NoReturn:
+        """Property method for get current
+        server transaction isolation level."""
 
-            if second_part:
-                for key in ("query", "query_src"):
-                    if key in kwargs:
-                        kwargs[key] = second_part.pop(0)
-                        break
-
-            self.logger.info(
-                f"Execute stream {part}/{total_parts} [native mode]"
-            )
-            output = dump_method(*args, **kwargs)
-
-            if second_part:
-                for query in second_part:
-                    part += 1
-                    self.logger.info(f"Execute query {part}/{total_parts}")
-                    cursor.execute(query)
-
-            if output:
-                self.refresh()
-
-            collect()
-            return output
-
-        return wrapper
-
-    def query_formatter(self, query: str) -> str | None:
-        """Reformat query."""
-
-        if not query:
-            return
-        return sql_format(sql=query, strip_comments=True).strip().strip(";")
+        raise NativeDumperValueError(
+            "Clickhouse server don't have transaction isolation level.",
+        )
 
     @multiquery
-    def __read_dump(
+    def _read_dump(
         self,
         fileobj: BufferedWriter,
         query: str | None,
@@ -207,7 +188,7 @@ class NativeDumper:
             raise NativeDumperReadError(error)
 
     @multiquery
-    def __write_between(
+    def _write_between(
         self,
         table_dest: str,
         table_src: str | None,
@@ -227,17 +208,9 @@ class NativeDumper:
             raise NativeDumperValueError(error_message)
 
         if not dumper_src:
-            cursor = HTTPCursor(
-                connector=self.connector,
-                compression_method=self.compression_method,
-                logger=self.logger,
-                timeout=self.cursor.timeout,
-            )
+            cursor = self.cursor
             src_dbname = self.dbname
             src_version = self.version
-            self.logger.info(
-                f"Set new connection for host {self.connector.host}."
-            )
         elif dumper_src.__class__ is NativeDumper:
             cursor = dumper_src.cursor
             src_dbname = dumper_src.dbname
@@ -284,7 +257,7 @@ class NativeDumper:
         self.write_dump(stream, table_dest, cursor.compression_method)
 
     @multiquery
-    def __to_reader(
+    def _to_reader(
         self,
         query: str | None,
         table_name: str | None,
@@ -308,20 +281,6 @@ class NativeDumper:
             columns=make_columns(self.cursor.metadata(f"({query}\n)")),
         )
         return self.cursor.get_stream(query)
-
-    def read_dump(
-        self,
-        fileobj: BufferedWriter,
-        query: str | None = None,
-        table_name: str | None = None,
-    ) -> bool:
-        """Read Native dump from Clickhouse."""
-
-        return self.__read_dump(
-            fileobj=fileobj,
-            query=query,
-            table_name=table_name,
-        )
 
     def write_dump(
         self,
@@ -376,34 +335,6 @@ class NativeDumper:
             f"Write into {self.connector.host}.{table_name} done."
         )
         self.refresh()
-
-    def write_between(
-        self,
-        table_dest: str,
-        table_src: str | None = None,
-        query_src: str | None = None,
-        dumper_src: Union["NativeDumper", object] = None,
-    ) -> bool:
-        """Write between Clickhouse servers."""
-
-        return self.__write_between(
-            table_dest=table_dest,
-            table_src=table_src,
-            query_src=query_src,
-            dumper_src=dumper_src,
-        )
-
-    def to_reader(
-        self,
-        query: str | None = None,
-        table_name: str | None = None,
-    ) -> NativeReader:
-        """Get stream from Clickhouse as NativeReader object."""
-
-        return self.__to_reader(
-            query=query,
-            table_name=table_name,
-        )
 
     def from_rows(
         self,
@@ -460,54 +391,8 @@ class NativeDumper:
         )
         self.refresh()
 
-    def from_pandas(
-        self,
-        data_frame: PdFrame,
-        table_name: str,
-    ) -> None:
-        """Write from pandas.DataFrame into Clickhouse table."""
-
-        self.from_rows(
-            dtype_data=iter(data_frame.values),
-            table_name=table_name,
-            source=DBMetadata(
-                name="pandas",
-                version="DataFrame",
-                columns=OrderedDict(zip(
-                    data_frame.columns,
-                    [str(dtype) for dtype in data_frame.dtypes],
-                )),
-            )
-        )
-
-    def from_polars(
-        self,
-        data_frame: PlFrame,
-        table_name: str,
-    ) -> None:
-        """Write from polars.DataFrame into Clickhouse table."""
-
-        self.from_rows(
-            dtype_data=data_frame.iter_rows(),
-            table_name=table_name,
-            source=DBMetadata(
-                name="polars",
-                version="DataFrame",
-                columns=OrderedDict(zip(
-                    data_frame.columns,
-                    [str(dtype) for dtype in data_frame.dtypes],
-                )),
-            )
-        )
-
     def refresh(self) -> None:
         """Refresh session."""
 
         self.cursor.refresh()
         self.logger.info(f"Connection to host {self.connector.host} updated.")
-
-    def close(self) -> None:
-        """Close session."""
-
-        self.cursor.close()
-        self.logger.info(f"Connection to host {self.connector.host} closed.")